# backend/openapi-schema.yml
openapi: 3.0.3
info:
  title: 高校选课系统 API
  version: 1.0.0
  description: 基于 Django REST Framework 的高校选课系统接口文档。
paths:
...
  /api/courses:
    get:
      operationId: courses_list
      description: 查询课程库。支持关键字搜索；传入 `paginate=1` 时，返回真实分页结构。
      summary: 获取课程列表
      parameters:
      - in: query
        name: page
        schema:
          type: integer
        description: 页码，从 1 开始。仅在 `paginate=1` 时生效。
      - in: query
        name: page_size
        schema:
          type: integer
        description: 每页条数，默认 10，最大 100。仅在 `paginate=1` 时生效。
      - in: query
        name: paginate
        schema:
          type: string
        description: 传入 `1` 时启用分页返回结构。
      - in: query
        name: q
        schema:
          type: string
        description: 关键字模糊搜索。
      tags:
      - courses
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCourseListResponse'
          description: ''
    post:
      operationId: courses_create
      description: 管理员创建课程库中的课程基础信息。
      summary: 创建课程
      tags:
      - courses
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Course'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Course'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
          description: ''
  /api/courses/{course_id}:
    get:
      operationId: courses_detail
      description: 根据课程 ID 获取单个课程详情。
      summary: 获取课程详情
      parameters:
      - in: path
        name: course_id
        schema:
          type: integer
        required: true
      tags:
      - courses
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
          description: ''
    put:
      operationId: courses_update
      description: 管理员更新指定课程的基础信息。
      summary: 更新课程
      parameters:
      - in: path
        name: course_id
        schema:
          type: integer
        required: true
      tags:
      - courses
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Course'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Course'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
          description: ''
    delete:
      operationId: courses_destroy
      description: 管理员删除指定课程。
      summary: 删除课程
      parameters:
      - in: path
        name: course_id
        schema:
          type: integer
        required: true
      tags:
      - courses
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/sections:
    get:
      operationId: sections_list
      description: 查询开课班级。支持按学期、课程、教师和关键字筛选；传入 `paginate=1` 时，返回真实分页结构。
      summary: 获取开课班级列表
      parameters:
      - in: query
        name: course_id
        schema:
          type: integer
        description: 按课程 ID 筛选。
      - in: query
        name: page
        schema:
          type: integer
        description: 页码，从 1 开始。仅在 `paginate=1` 时生效。
      - in: query
        name: page_size
        schema:
          type: integer
        description: 每页条数，默认 10，最大 100。仅在 `paginate=1` 时生效。
      - in: query
        name: paginate
        schema:
          type: string
        description: 传入 `1` 时启用分页返回结构。
      - in: query
        name: q
        schema:
          type: string
        description: 关键字模糊搜索。
      - in: query
        name: teacher_id
        schema:
          type: integer
        description: 按教师用户 ID 筛选。
      - in: query
        name: term_id
        schema:
          type: integer
        description: 按学期 ID 筛选。
      tags:
      - sections
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSectionListResponse'
          description: ''
    post:
      operationId: sections_create
      description: 管理员为课程创建具体开课班级。
      summary: 创建开课班级
      tags:
      - sections
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Section'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Section'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Section'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
          description: ''
  /api/sections/{section_id}:
    get:
      operationId: sections_detail
      description: 根据开课班级 ID 获取详情。
      summary: 获取开课班级详情
      parameters:
      - in: path
        name: section_id
        schema:
          type: integer
        required: true
      tags:
      - sections
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
          description: ''
    put:
      operationId: sections_update
      description: 管理员更新指定开课班级信息。
      summary: 更新开课班级
      parameters:
      - in: path
        name: section_id
        schema:
          type: integer
        required: true
      tags:
      - sections
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Section'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Section'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Section'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
          description: ''
    delete:
      operationId: sections_destroy
      description: 管理员删除指定开课班级。
      summary: 删除开课班级
      parameters:
      - in: path
        name: section_id
        schema:
          type: integer
        required: true
      tags:
      - sections
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        code:
          type: string
          maxLength: 50
        name:
          type: string
          maxLength: 200
        credits:
          type: integer
          maximum: 4294967295
          minimum: 0
          format: int64
        description:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - code
      - created_at
      - id
      - name
      - updated_at
...
    PaginatedCourseListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Course'
        count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
      required:
      - count
      - page
      - page_size
      - results
      - total_pages
...
    PaginatedSectionListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Section'
        count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
      required:
      - count
      - page
      - page_size
      - results
      - total_pages
...
    Section:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        term:
          type: integer
        term_name:
          type: string
          readOnly: true
        course:
          type: integer
        course_name:
          type: string
          readOnly: true
        course_code:
          type: string
          readOnly: true
        teacher:
          type: integer
        teacher_name:
          type: string
          readOnly: true
        teacher_username:
          type: string
          readOnly: true
        name:
          type: string
          maxLength: 100
        capacity:
          type: integer
          maximum: 4294967295
          minimum: 0
          format: int64
        weekday:
          allOf:
          - $ref: '#/components/schemas/WeekdayEnum'
          minimum: 0
          maximum: 65535
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        location:
          type: string
          maxLength: 100
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - course
      - course_code
      - course_name
      - created_at
      - end_time
      - id
      - name
      - start_time
      - teacher
      - teacher_name
      - teacher_username
      - term
      - term_name
      - updated_at
      - weekday
...
    WeekdayEnum:
      enum:
      - 1
      - 2
      - 3
      - 4
      - 5
      - 6
      - 7
      type: integer
      description: |-
        * `1` - Monday
        * `2` - Tuesday
        * `3` - Wednesday
        * `4` - Thursday
        * `5` - Friday
        * `6` - Saturday
        * `7` - Sunday
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT